Remove metadata entirely.
Turns out the only place this was used was to pass back to functions or layers that the
user would be specifying themselves anyway -- so they can punch the metadata around. I
think this leads to a simpler API where I don't say anything about type/color/merging
behavior/whatever, and that can be built out of these primitives or ad-hoc by the
consumer.