Fix an issue that could cause the Y domain to grow or shrink forever.
The Y domain selector respects any overloads coming through ChartProvider props.
DataLoaders are the only possible source of truth for the true Y domain of a particular
batch of data, so if they hit a fallback clause and return one of the Y domains they
were given, that truthiness is then lost. If the consumer additionally does some
post-processing on the Y domains they receive, such as by extending it on either end
for display purposes, that transform may be applied to itself repeatedly as the
fallback clause returns the already-transformed Y domain, causing it to grow forever.