Don't issue (expensive!) loads in some situations.
Often, series IDs and the X domain will be recomputed on the fly but end up
value-equals. In a normal React library, this wouldn't cause tons of expensive
loads or rendering, but here it does. This commit cheats a little bit on the
"pure render" contract by also deep-checking some of the more common things
that change, that can be deep checked, that will issue a load. In particular,
this includes the X domain (internal and override), the list of series IDs
and the physical chart size. Notably absent from this list is the data loader,
which can do a shallow check, but as it is a function it cannot do a
meaningful deep check.