Fix BucketedLineLayer line rendering. Fixes #55.
This is an issue with the differing ways that rects and lines handle
coordinates. Say you have a 1x1 pixel view, and you draw a rectangle
from (0, 0) to (1, 1):

      1
   ___  1
  |   |
  |   |
0  ‾‾‾
  0

If you draw a line starting from (1, 1), that will start from here:

   ___.
  |   |
  |   |
   ‾‾‾

Makes sense, but this causes blurry lines. For that reason, the canvas
is translated half a pixel in each dimension so you start the line here:

    .
   ___
  |   |
  |   |
   ‾‾‾

Which gives you pixel-perfect lines, except that they start outside the
bounds of the rectangle. Same would happen if we translated in the other
direction.

The most correct fix would be something along the lines of scaling the
value down from a valid range of (height) to (height - 1) before flooring
it, but that's a lot of math that I'm not even sure is right for a
1-pixel correction, so just clamp it and be done with it.